feat: Add generateKey function and update package version
A new utility function `generateKey` has been added which enables to generate keys based on the provided hash and encoding options. The `generateKey` function is also exported in the main index file. Additionally, a minor change has been made in the caching logic with an added log method and the package version has been revised to 1.1.1.