docs: Update documentation files and links
Several changes have been made to the documentation files. A syntax error was fixed in the app.config.ts file, where a trailing comma was removed from the github link. The "Getting Started" section of the documentation was updated with additional information and a direct link to the npm package. Finally, a faulty GitHub link was replaced in the content index file.