feat: Add 'files' method to FileSysCache class
This update introduces a new 'files' method in the FileSysCache class. The 'files' method returns a list of files in the cache including their attributes such as name, size, ttl, expiration, and 'expires_in'. Unit tests and documentation for this new feature have been also added. The package version has been bumped from 2.0.1 to 2.1.0 to reflect this addition.